+++ src/openapi2excel.tests/TranslationsTest.cs
@@ -88,6 +88,7 @@ namespace OpenApi2Excel.Tests
          Assert.Contains("{1}", translation.BodyFormatWithObject);
          Assert.Contains("{0}", translation.InfoOperationWithId);
          Assert.Contains("{1}", translation.InfoOperationWithId);
+         Assert.Contains("{0}", translation.AllOfType);
       }
 
       [Theory]